Opt and vis edits
Add another unit of habitat gained by removing barrier 8 so that the hydrography constriant is binding. You want to remove 4 of: 8 (for 4), 1 (for 4), 10 (for 4), and 4 (for 4), but 8, 4, and 10 all require downstream removal. Also get rid of the "known solutions" because they are wrong. In fact, for div and eq you want to remove less than 4 (which I didn't consider).